pyproject.toml
[tool.poetry]
name = "imagetiffcollector"
version = "0.1.0"
description = ""
authors = ["Vladimir Titov <tvv-vg@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
